Properly parse device commissioning response.
It turns out that one byte of data is getting reliably dropped by
the adapter (0x13). So this won't work with the actual device, but 
without the missing byte the checksum returned by the device is invalid
(and specs fail). So while this is change is technically correct it
won't work until I can determine why it's dropping that byte.