Switch to libftdi-ruby from serialport.
Previously hacklet relied on a kernel driver to communicate with the
dongle. This change switches it to a cross platform library, which
allows hacklet to run on OSX as well as Windows.

Note: The Windows support is unconfirmed but should theoretically work.
I leave it as an exercise for a intrepid Windows user to confirm.